Refactor fingerprint table
