+++ frontend/src/pages/scenes/Scene.tsx
@@ -1,21 +1,13 @@
-import { FC, useContext, useMemo } from "react";
+import { FC, useContext } from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
-import { Button, Card, Tabs, Tab, Table } from "react-bootstrap";
-import {
-  faCheckCircle,
-  faTimesCircle,
-  faSpinner,
-  faTriangleExclamation,
-} from "@fortawesome/free-solid-svg-icons";
+import { Button, Card, Tabs, Tab } from "react-bootstrap";
 
 import {
   usePendingEditsCount,
   TargetTypeEnum,
-  useUnmatchFingerprint,
   SceneFragment as Scene,
 } from "src/graphql";
 import AuthContext from "src/AuthContext";
-import { useToast } from "src/hooks";
 import {
   canEdit,
   tagHref,
@@ -23,26 +15,15 @@ import {
   studioHref,
   createHref,
   formatDuration,
-  formatDateTime,
   formatPendingEdits,
   getUrlBySite,
   compareByName,
 } from "src/utils";
-import {
-  ROUTE_SCENE_EDIT,
-  ROUTE_SCENES,
-  ROUTE_SCENE_DELETE,
-} from "src/constants/route";
-import {
-  GenderIcon,
-  TagLink,
-  PerformerName,
-  Icon,
-} from "src/components/fragments";
+import { ROUTE_SCENE_EDIT, ROUTE_SCENE_DELETE } from "src/constants/route";
+import { GenderIcon, TagLink, PerformerName } from "src/components/fragments";
 import { EditList, URLList } from "src/components/list";
 import Image from "src/components/image";
-
-type Fingerprint = NonNullable<Scene["fingerprints"][number]>;
+import { FingerprintTable } from "./components/fingerprints";
 
 const DEFAULT_TAB = "description";
 
@@ -55,9 +36,6 @@ const SceneComponent: FC<Props> = ({ scene }) => {
   const location = useLocation();
   const activeTab = location.hash?.slice(1) || DEFAULT_TAB;
   const auth = useContext(AuthContext);
-  const addToast = useToast();
-
-  const [unmatchFingerprint, { loading: unmatching }] = useUnmatchFingerprint();
 
   const { data: editData } = usePendingEditsCount({
     type: TargetTypeEnum.SCENE,
@@ -83,92 +61,6 @@ const SceneComponent: FC<Props> = ({ scene }) => {
       );
     })
     .map((p, index) => (index % 2 === 2 ? [" • ", p] : p));
-
-  const fingerprints = useMemo(() => {
-    async function handleFingerprintUnmatch(fingerprint: Fingerprint) {
-      if (unmatching) return;
-
-      const { data } = await unmatchFingerprint({
-        variables: {
-          scene_id: scene.id,
-          algorithm: fingerprint.algorithm,
-          hash: fingerprint.hash,
-          duration: fingerprint.duration,
-        },
-      });
-      const success = data?.unmatchFingerprint;
-      addToast({
-        variant: success ? "success" : "danger",
-        content: `${
-          success ? "Removed" : "Failed to remove"
-        } fingerprint submission`,
-      });
-    }
-
-    function maybeRenderSubmitted(fingerprint: Fingerprint, reported: boolean) {
-      if (
-        (!reported && fingerprint.user_submitted) ||
-        (reported && fingerprint.user_reported)
-      ) {
-        return (
-          <Button
-            className="user-submitted"
-            title="Submitted by you - click to remove submission"
-            onKeyDown={() => handleFingerprintUnmatch(fingerprint)}
-            onClick={() => handleFingerprintUnmatch(fingerprint)}
-            variant="link"
-          >
-            {!unmatching ? (
-              <>
-                <Icon icon={faCheckCircle} />
-                <Icon icon={faTimesCircle} />
-              </>
-            ) : (
-              <Icon icon={faSpinner} className="fa-spin" />
-            )}
-          </Button>
-        );
-      }
-    }
-
-    function renderSubmissions(fingerprint: Fingerprint) {
-      return (
-        <span className="fingerprint-submissions">
-          <span className="fingerprint-submissions-plus">
-            {fingerprint.submissions}
-          </span>
-          {fingerprint.reports > 0 && (
-            <span className="fingerprint-reports">
-              {fingerprint.reports} <Icon icon={faTriangleExclamation} />
-              {maybeRenderSubmitted(fingerprint, true)}
-            </span>
-          )}
-          {maybeRenderSubmitted(fingerprint, false)}
-        </span>
-      );
-    }
-
-    return scene.fingerprints.map((fingerprint) => (
-      <tr key={fingerprint.hash}>
-        <td>{fingerprint.algorithm}</td>
-        <td className="font-monospace">
-          <Link
-            to={`${createHref(ROUTE_SCENES)}?fingerprint=${fingerprint.hash}`}
-          >
-            {fingerprint.hash}
-          </Link>
-        </td>
-        <td>
-          <span title={`${fingerprint.duration}s`}>
-            {formatDuration(fingerprint.duration)}
-          </span>
-        </td>
-        <td>{renderSubmissions(fingerprint)}</td>
-        <td>{formatDateTime(fingerprint.created)}</td>
-        <td>{formatDateTime(fingerprint.updated)}</td>
-      </tr>
-    ));
-  }, [scene.fingerprints, unmatchFingerprint, unmatching, addToast, scene.id]);
 
   const tags = [...scene.tags].sort(compareByName).map((tag) => (
     <li key={tag.name}>
@@ -283,38 +175,7 @@ const SceneComponent: FC<Props> = ({ scene }) => {
           )}
         </Tab>
         <Tab eventKey="fingerprints" title="Fingerprints" mountOnEnter={false}>
-          <div className="scene-fingerprints my-4">
-            <h4>Fingerprints:</h4>
-            {fingerprints.length === 0 ? (
-              <h6>No fingerprints found for this scene.</h6>
-            ) : (
-              <Table striped variant="dark">
-                <thead>
-                  <tr>
-                    <td>
-                      <b>Algorithm</b>
-                    </td>
-                    <td>
-                      <b>Hash</b>
-                    </td>
-                    <td>
-                      <b>Duration</b>
-                    </td>
-                    <td>
-                      <b>Submissions</b>
-                    </td>
-                    <td>
-                      <b>First Added</b>
-                    </td>
-                    <td>
-                      <b>Last Added</b>
-                    </td>
-                  </tr>
-                </thead>
-                <tbody>{fingerprints}</tbody>
-              </Table>
-            )}
-          </div>
+          <FingerprintTable scene={scene} />
         </Tab>
         <Tab eventKey="links" title="Links">
           <URLList urls={scene.urls} />
